+++ webapp/src/components/AssetPage/CategoryBadge/CategoryBadge.tsx
