+++ webapp/src/components/AssetPage/CategoryBadge/CategoryBadge.types.ts
@@ -1,5 +1,5 @@
 import { NFT } from '@dcl/schemas'
-import { AssetType } from '../../../../modules/asset/types'
+import { AssetType } from '../../../modules/asset/types'
 
 export type Props = {
   wearable: Required<NFT['data']>['wearable']
