+++ webapp/src/components/AssetPage/CategoryBadge/index.ts
