+++ webapp/src/components/AssetPage/Collection/Collection.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import { locations } from '../../../../modules/routing/locations'
+import { locations } from '../../../modules/routing/locations'
 import { Props } from './Collection.types'
-import CollectionImage from '../../../CollectionImage'
+import CollectionImage from '../../CollectionImage'
 import { Stats } from 'decentraland-ui'
-import CollectionProvider from '../../../CollectionProvider'
+import CollectionProvider from '../../CollectionProvider'
 import styles from './Collection.module.css'
 
 const Collection = (props: Props) => {
