+++ webapp/src/components/AssetPage/Collection/Collection.types.tsx
@@ -1,0 +1,5 @@
+import { Asset } from '../../../modules/asset/types'
+
+export type Props = {
+  asset: Asset
+}
