+++ webapp/src/components/AssetPage/ENSDetail/ENSDetail.tsx
@@ -4,12 +4,12 @@ import { AssetImage } from '../../AssetImage'
 import { PageHeader } from '../../PageHeader'
 import { Network } from '../Network'
 import { Props } from './ENSDetail.types'
-import Title from '../V2/Title'
+import Title from '../Title'
 import { Box } from '../../AssetBrowse/Box'
-import { Owner } from '../V2/Owner'
+import { Owner } from '../Owner'
 import ListedBadge from '../../ListedBadge'
-import Price from '../V2/Price'
-import Expiration from '../V2/Expiration'
+import Price from '../Price'
+import Expiration from '../Expiration'
 import { Actions } from '../Actions'
 import { Bids } from '../Bids'
 import { TransactionHistory } from '../TransactionHistory'
