+++ webapp/src/components/AssetPage/Expiration/Expiration.container.ts
@@ -1,8 +1,8 @@
 import { connect } from 'react-redux'
 import { MapStateProps } from './Expiration.types'
 import Expiration from './Expiration'
-import { RootState } from '../../../../modules/reducer'
-import { getCurrentOrder } from '../../../../modules/order/selectors'
+import { RootState } from '../../../modules/reducer'
+import { getCurrentOrder } from '../../../modules/order/selectors'
 
 const mapState = (state: RootState): MapStateProps => ({
   order: getCurrentOrder(state) || undefined
