+++ webapp/src/components/AssetPage/Expiration/Expiration.types.ts
@@ -1,4 +1,4 @@
-import { Order } from '../../../../modules/order/types'
+import { Order } from '../../../modules/order/types'
 
 export type Props = {
   order?: Order
