+++ webapp/src/components/AssetPage/GenderBadge/GenderBadge.container.ts
@@ -2,9 +2,9 @@ import { BodyShape, WearableGender } from '@dcl/schemas'
 import { push } from 'connected-react-router'
 import { connect } from 'react-redux'
 import { Dispatch } from 'redux'
-import { isGender, isUnisex } from '../../../../modules/nft/wearable/utils'
-import { locations } from '../../../../modules/routing/locations'
-import { Section } from '../../../../modules/vendor/decentraland'
+import { isGender, isUnisex } from '../../../modules/nft/wearable/utils'
+import { locations } from '../../../modules/routing/locations'
+import { Section } from '../../../modules/vendor/decentraland'
 import GenderBadge from './GenderBadge'
 import { MapDispatchProps, OwnProps } from './GenderBadge.types'
