+++ webapp/src/components/AssetPage/IconBadge/IconBadge.tsx
