+++ webapp/src/components/AssetPage/IconBadge/IconBadge.types.ts
