+++ webapp/src/components/AssetPage/IconBadge/index.ts
