+++ webapp/src/components/AssetPage/Owner/Owner.tsx
@@ -3,10 +3,10 @@ import { Link } from 'react-router-dom'
 import { Profile } from 'decentraland-dapps/dist/containers'
 import { t } from 'decentraland-dapps/dist/modules/translation/utils'
 import { Stats } from 'decentraland-ui'
-import { locations } from '../../../../modules/routing/locations'
-import { BrowseOptions } from '../../../../modules/routing/types'
-import { Section } from '../../../../modules/vendor/decentraland'
-import { AssetType } from '../../../../modules/asset/types'
+import { locations } from '../../../modules/routing/locations'
+import { BrowseOptions } from '../../../modules/routing/types'
+import { Section } from '../../../modules/vendor/decentraland'
+import { AssetType } from '../../../modules/asset/types'
 import { Props } from './Owner.types'
 
 import './Owner.css'
