+++ webapp/src/components/AssetPage/Price/Price.container.ts
@@ -1,8 +1,8 @@
 import { connect } from 'react-redux'
 import { MapStateProps, OwnProps } from './Price.types'
 import Price from './Price'
-import { RootState } from '../../../../modules/reducer'
-import { getCurrentOrder } from '../../../../modules/order/selectors'
+import { RootState } from '../../../modules/reducer'
+import { getCurrentOrder } from '../../../modules/order/selectors'
 import { Item } from '@dcl/schemas'
 
 const mapState = (state: RootState, { asset }: OwnProps): MapStateProps => {
