+++ webapp/src/components/AssetPage/RarityBadge/RarityBadge.container.ts
@@ -2,8 +2,8 @@ import { push } from 'connected-react-router'
 import { connect } from 'react-redux'
 import { Dispatch } from 'redux'
 import RarityBadge from './RarityBadge'
-import { locations } from '../../../../modules/routing/locations'
-import { Section } from '../../../../modules/vendor/decentraland'
+import { locations } from '../../../modules/routing/locations'
+import { Section } from '../../../modules/vendor/decentraland'
 import { MapDispatchProps, OwnProps } from './RarityBadge.types'
 
 const mapDispatch = (
