+++ webapp/src/components/AssetPage/RarityBadge/RarityBadge.module.css
