+++ webapp/src/components/AssetPage/RarityBadge/RarityBadge.tsx
