+++ webapp/src/components/AssetPage/RarityBadge/RarityBadge.types.ts
@@ -1,5 +1,5 @@
 import { Rarity } from '@dcl/schemas'
-import { AssetType } from '../../../../modules/asset/types'
+import { AssetType } from '../../../modules/asset/types'
 
 export type Props = {
   rarity: Rarity
