+++ webapp/src/components/AssetPage/RarityBadge/index.ts
