+++ webapp/src/components/AssetPage/Title/Title.types.ts
@@ -1,0 +1,5 @@
+import { Asset } from '../../../modules/asset/types'
+
+export type Props = {
+  asset: Asset
+}
