--- webapp/src/components/AssetPage/V2/Collection/Collection.types.tsx
@@ -1,5 +1,0 @@
-import { Asset } from '../../../../modules/asset/types'
-
-export type Props = {
-  asset: Asset
-}
