--- webapp/src/components/AssetPage/V2/IconBadge/IconBadge.css
@@ -1,75 +1,0 @@
-.IconBadge .BaseMale {
-  background-image: url(../../../../images/wearables/MaleIcon.svg);
-}
-
-.IconBadge .BaseFemale {
-  background-image: url(../../../../images/wearables/FemaleIcon.svg);
-}
-
-.IconBadge .Unisex {
-  background-image: url(../../../../images/wearables/UnisexIcon.svg);
-}
-
-.IconBadge .body_shape {
-  background-image: url(../../../../images/wearables/BodyShapeIcon.svg);
-}
-
-.IconBadge .earring {
-  background-image: url(../../../../images/wearables/EarringsIcon.svg);
-}
-
-.IconBadge .eyebrows {
-  background-image: url(../../../../images/wearables/EyebrowIcon.svg);
-}
-
-.IconBadge .eyes {
-  background-image: url(../../../../images/wearables/EyesIcon.svg);
-}
-
-.IconBadge .eyewear {
-  background-image: url(../../../../images/wearables/EyewearIcon.svg);
-}
-
-.IconBadge .facial_hair {
-  background-image: url(../../../../images/wearables/FacilHairIcon.svg);
-}
-
-.IconBadge .feet {
-  background-image: url(../../../../images/wearables/FeetIcon.svg);
-}
-
-.IconBadge .hair {
-  background-image: url(../../../../images/wearables/HairIcon.svg);
-}
-
-.IconBadge .hat {
-  background-image: url(../../../../images/wearables/HatIcon.svg);
-}
-
-.IconBadge .helmet {
-  background-image: url(../../../../images/wearables/HelmetIcon.svg);
-}
-
-.IconBadge .lower_body {
-  background-image: url(../../../../images/wearables/LowerBodyIcon.svg);
-}
-
-.IconBadge .mask {
-  background-image: url(../../../../images/wearables/MaskIcon.svg);
-}
-
-.IconBadge .mouth {
-  background-image: url(../../../../images/wearables/MouthIcon.svg);
-}
-
-.IconBadge .tiara {
-  background-image: url(../../../../images/wearables/TiaraIcon.svg);
-}
-
-.IconBadge .top_head {
-  background-image: url(../../../../images/wearables/TopHeadIcon.svg);
-}
-
-.IconBadge .upper_body {
-  background-image: url(../../../../images/wearables/UpperBodyIcon.svg);
-}
