--- webapp/src/components/AssetPage/V2/Title/Title.types.ts
@@ -1,5 +1,0 @@
-import { Asset } from '../../../../modules/asset/types'
-
-export type Props = {
-  asset: Asset
-}
