+++ webapp/src/components/AssetPage/WearableDetail/WearableDetail.tsx
@@ -7,18 +7,18 @@ import { PageHeader } from '../../PageHeader'
 import { Network } from '../Network'
 import { Description } from '../Description'
 import { Props } from './WearableDetail.types'
-import Title from '../V2/Title'
-import RarityBadge from '../V2/RarityBadge'
+import Title from '../Title'
+import RarityBadge from '../RarityBadge'
 import { AssetType } from '../../../modules/asset/types'
-import GenderBadge from '../V2/GenderBadge'
-import CategoryBadge from '../V2/CategoryBadge'
+import GenderBadge from '../GenderBadge'
+import CategoryBadge from '../CategoryBadge'
 import { Box } from '../../AssetBrowse/Box'
-import { Owner } from '../V2/Owner'
-import Collection from '../V2/Collection'
+import { Owner } from '../Owner'
+import Collection from '../Collection'
 import ListedBadge from '../../ListedBadge'
 import styles from './WearableDetail.module.css'
-import Price from '../V2/Price'
-import Expiration from '../V2/Expiration'
+import Price from '../Price'
+import Expiration from '../Expiration'
 import { Actions } from '../Actions'
 import { Bids } from '../Bids'
 import { TransactionHistory } from '../TransactionHistory'
